Support both ESM and CDN, move code to /src/
