+++ .replit
@@ -27,7 +27,7 @@ requiredFiles = [".replit", "replit.nix", ".config"]
     start = "typescript-language-server --stdio"
 
 [deployment]
-run = ["tsx", "index.ts"]
+run = ["tsx", "src/index.ts"]
 deploymentTarget = "cloudrun"
 ignorePorts = false
