+++ src/module.ts
@@ -1,0 +1,4 @@
+import {Quaternion, Vector3, PerspectiveCamera} from "three";
+import Three3DCamera from "./index";
+
+export default Three3DCamera(Quaternion, Vector3, PerspectiveCamera)
